+++ tests/test_cli.py
@@ -1,4 +1,9 @@
 import os
+import shutil
+import uuid
+from pathlib import Path
+from py import path as pypath
+
 from laktory import app
 from laktory import settings
 from laktory import models
@@ -10,145 +15,183 @@ settings.cli_raise_external_exceptions = True
 paths = Paths(__file__)
 
 
-def test_preview_pulumi():
-    filepath = os.path.join(paths.data, "stack.yaml")
-    result = runner.invoke(app, ["preview", "--env", "dev", "--filepath", filepath])
-    assert result.exit_code == 0
+def _read_stack(template, backend):
 
-
-def test_preview_terraform():
-    filepath = os.path.join(paths.data, "stack.yaml")
-
-    # Ideally, we would run `laktory init`, but the runner does not seem to handle running multiple commands
-    with open(filepath, "r") as fp:
-        pstack = models.Stack.model_validate_yaml(fp).to_terraform(env_name="dev")
-        pstack.init(flags=["-migrate-state", "-upgrade"])
-
-    result = runner.invoke(
-        app,
-        ["preview", "--backend", "terraform", "--env", "dev", "--filepath", filepath],
+    dirpath = pypath.local(
+        f"{paths.tmp}/quickstart_{template}_{backend}_{str(uuid.uuid4())}"
     )
-    assert result.exit_code == 0
+    stack_filepath = dirpath / "stack.yaml"
 
+    # Change dir
+    os.mkdir(dirpath)
 
-def test_quickstart_pulumi():
-    filepath = os.path.join(paths.tmp, "stack_quickstart_pulumi.yaml")
-    print(filepath)
-    result = runner.invoke(
-        app,
-        [
-            "quickstart",
-            "--backend",
-            "pulumi",
-            "-o",
-            "okube",
-            "-n",
-            "Standard_DS3_v2",
-            "--filepath",
-            filepath,
-        ],
-    )
-
-    # TODO: Add validation for the generated stack?
-    with open(filepath) as fp:
-        stack = models.Stack.model_validate_yaml(fp)
-
-    data = stack.model_dump(exclude_unset=True)
-
-    assert stack.backend == "pulumi"
-    assert stack.name == "quickstart"
-    assert stack.pulumi.config["databricks:host"] == "${vars.DATABRICKS_HOST}"
-    assert stack.pulumi.config["databricks:token"] == "${vars.DATABRICKS_TOKEN}"
-    assert len(stack.resources.databricks_dbfsfiles) == 1
-    assert len(stack.resources.databricks_notebooks) == 1
-    assert len(stack.resources.pipelines) == 1
-
-
-def test_quickstart_terraform():
-    filepath = os.path.join(paths.tmp, "stack_quickstart_terraform.yaml")
-    result = runner.invoke(
-        app,
-        [
-            "quickstart",
-            "--backend",
-            "terraform",
-            "-o",
-            "okube",
-            "-n",
-            "Standard_DS3_v2",
-            "--filepath",
-            filepath,
-        ],
-    )
-
-    with open(filepath) as fp:
-        stack = models.Stack.model_validate_yaml(fp)
-
-    data = stack.model_dump(exclude_unset=True)
-
-    assert stack.backend == "terraform"
-    assert stack.name == "quickstart"
-    assert stack.resources.providers["databricks"].host == "${vars.DATABRICKS_HOST}"
-    assert stack.resources.providers["databricks"].token == "${vars.DATABRICKS_TOKEN}"
-    assert len(stack.resources.databricks_dbfsfiles) == 1
-    assert len(stack.resources.databricks_notebooks) == 1
-    assert len(stack.resources.pipelines) == 1
-
-
-def atest_deploy_pulumi():
-    # TODO: Figure out how to run in isolation. Currently, pulumi up commands
-    # are run concurrently because of the multiple python testing environment
-    # which result in:  Conflict: Another update is currently in progress
-    for filename in [
-        "stack.yaml",
-        "stack_empty.yaml",
-    ]:
-        filepath = os.path.join(paths.data, filename)
-        result = runner.invoke(
+    # Copy Stack
+    with dirpath.as_cwd():
+        _ = runner.invoke(
             app,
-            [
-                "deploy",
-                "-e",
-                "dev",
-                "--filepath",
-                filepath,
-                "--pulumi-options",
-                "--yes",
-            ],
+            ["quickstart", "--template", template, "--backend", backend],
         )
-        assert result.exit_code == 0
+
+    # Read Stack
+    with open(stack_filepath) as fp:
+        stack = models.Stack.model_validate_yaml(fp)
+    data = stack.model_dump()
+
+    # Count resources
+    resources_count = 0
+    for k in data["resources"]:
+        if k == "providers":
+            continue
+        resources_count += len(data["resources"][k].keys())
+
+    if template == "workspace":
+        assert resources_count == 7
+        assert (
+            "directory-laktory-dashboards"
+            in data["resources"]["databricks_directories"].keys()
+        )
+        assert (
+            "secret-scope-laktory"
+            in data["resources"]["databricks_secretscopes"].keys()
+        )
+        assert "warehouse-laktory" in data["resources"]["databricks_warehouses"].keys()
+        pass
+    elif template == "workflows":
+        assert resources_count == 7
+        assert (
+            "dbfs-file-stock-prices" in data["resources"]["databricks_dbfsfiles"].keys()
+        )
+        assert (
+            "notebook-job-laktory-pl"
+            in data["resources"]["databricks_notebooks"].keys()
+        )
+        assert (
+            "notebook-dlt-laktory-pl"
+            in data["resources"]["databricks_notebooks"].keys()
+        )
+        assert "job-hello" in data["resources"]["databricks_jobs"].keys()
+        assert "pl-stocks-sql" in data["resources"]["pipelines"].keys()
+        assert "pl-stocks-spark-dlt" in data["resources"]["pipelines"].keys()
+    elif template == "unity-catalog":
+        assert resources_count == 8
+        assert "group-laktory-friends" in data["resources"]["databricks_groups"].keys()
+        assert "user-kubic-okube-ai" in data["resources"]["databricks_users"].keys()
+        assert "catalog-lqs-prd" in data["resources"]["databricks_catalogs"].keys()
+
+    else:
+        raise ValueError()
+
+    # Cleanup
+    shutil.rmtree(dirpath)
 
 
-def atest_deploy_terraform():
-    # TODO: Figure out how to run in isolation. Currently, pulumi up commands
-    # are run concurrently because of the multiple python testing environment
-    # which result in:  Conflict: Another update is currently in progress
-    for filename in [
-        "stack.yaml",
-        "stack_empty.yaml",
-    ]:
-        filepath = os.path.join(paths.data, filename)
-        result = runner.invoke(
+def _preview_stack(template, backend, env):
+
+    dirpath = pypath.local(
+        f"{paths.tmp}/quickstart_{template}_{backend}_{str(uuid.uuid4())}"
+    )
+    stack_filepath = dirpath / "stack.yaml"
+
+    # Change dir
+    os.mkdir(dirpath)
+
+    # Copy Stack
+    with dirpath.as_cwd():
+        _ = runner.invoke(
             app,
-            [
-                "deploy",
-                "-e",
-                "dev",
-                "--backend",
-                "terraform",
-                "--filepath",
-                filepath,
-                "--terraform-options",
-                "--auto-approve",
-            ],
+            ["quickstart", "--template", template, "--backend", backend],
         )
-        assert result.exit_code == 0
+
+    if backend == "terraform":
+        # Ideally, we would run `laktory init`, but the runner does not seem to handle running multiple commands
+        with open(stack_filepath, "r") as fp:
+            pstack = models.Stack.model_validate_yaml(fp).to_terraform(env_name=env)
+            pstack.init(flags=["-migrate-state", "-upgrade"])
+
+    # Preview Stack
+    cmds = ["preview"]
+    if env:
+        cmds += ["--env", env]
+    _ = runner.invoke(app, cmds)
+
+    # Cleanup
+    shutil.rmtree(dirpath)
+
+
+def _deploy_stack(template, backend, env):
+
+    dirpath = pypath.local(
+        f"{paths.tmp}/quickstart_{template}_{backend}_{str(uuid.uuid4())}"
+    )
+    stack_filepath = dirpath / "stack.yaml"
+
+    # Change dir
+    os.mkdir(dirpath)
+
+    # Copy Stack
+    with dirpath.as_cwd():
+        _ = runner.invoke(
+            app,
+            ["quickstart", "--template", template, "--backend", backend],
+        )
+
+    if backend == "terraform":
+        # Ideally, we would run `laktory init`, but the runner does not seem to handle running multiple commands
+        with open(stack_filepath, "r") as fp:
+            pstack = models.Stack.model_validate_yaml(fp).to_terraform(env_name=env)
+            pstack.init(flags=["-migrate-state", "-upgrade"])
+
+    # Deploy Stack
+    cmds = ["deploy"]
+    if env:
+        cmds += ["-e", env]
+    if backend == "pulumi":
+        cmds += ["--pulumi-options", "--yes"]
+    elif backend == "terraform":
+        cmds += ["--terraform-options", "--auto-approve"]
+    _ = runner.invoke(app, cmds)
+
+    # Cleanup
+    shutil.rmtree(dirpath)
+
+
+def test_read_quickstart_stacks():
+    for template, backend in [
+        ("workspace", "terraform"),
+        ("workspace", "pulumi"),
+        ("workflows", "terraform"),
+        ("workflows", "pulumi"),
+        ("unity-catalog", "terraform"),
+        ("unity-catalog", "pulumi"),
+    ]:
+        _read_stack(template, backend)
+
+
+def test_preview_quickstart_stacks():
+    for template, backend, env in [
+        ("workspace", "terraform", "dev"),
+        ("workspace", "pulumi", "dev"),
+        ("workflows", "terraform", "dev"),
+        ("workflows", "pulumi", "dev"),
+        ("unity-catalog", "terraform", None),
+        ("unity-catalog", "pulumi", None),
+    ]:
+        _preview_stack(template, backend, env)
+
+
+def atest_deploy_quickstart_stacks():
+    for template, backend, env in [
+        ("workspace", "terraform", "dev"),
+        ("workspace", "pulumi", "dev"),
+        ("workflows", "terraform", "dev"),
+        ("workflows", "pulumi", "dev"),
+        ("unity-catalog", "terraform"),
+        ("unity-catalog", "pulumi"),
+    ]:
+        _deploy_stack(template, backend, env)
 
 
 if __name__ == "__main__":
-    test_preview_pulumi()
-    test_preview_terraform()
-    test_quickstart_pulumi()
-    test_quickstart_terraform()
-    # atest_deploy_pulumi()
-    # atest_deploy_terraform()
+    test_read_quickstart_stacks()
+    test_preview_quickstart_stacks()
+    # atest_deploy_quickstart_stacks()