+++ CHANGELOG.md
@@ -4,8 +4,13 @@
 ### Added
 * Support for multi-segments (semi-column ; separated) SQL statements 
 * Support for Databricks Lakeview Dashboard resource
+* CLI `destroy` command
+* `unity-catalog`, `workspace` and `workflows` template choices for CLI `quickstart`
 ### Updated
 * Better feedback when terraform is not installed
+### Breaking changes
+* Removed `backend` and `organization` arguments for CLI
+* Combined CLI `pulumi-options` and `terraform-options` into `options`
 
 ## [0.4.11] - 2024-08-16
 ### Added